Implemented cancels requests and interrupts the execution in the kernel; Refactored
